requirements.txt
streamlit
pandas
requests
selenium
plotly
nltk
google-generativeai
pymongo
lxml
webdriver-manager
kafka-python
python-dotenv
wordcloud
